added new npm modules for firebase, http and fs to get tweets pushed to firebase DB and get started with live reload server
