+++ assets/js/app.js
@@ -1,16 +1,3 @@
-// Initialize Firebase
-var config = {
-    apiKey: "AIzaSyB_zXnNbvnEYRsYGSpumtV-jiADhtmvhtM",
-    authDomain: "utcodingbootcampproject1.firebaseapp.com",
-    databaseURL: "https://utcodingbootcampproject1.firebaseio.com",
-    projectId: "utcodingbootcampproject1",
-    storageBucket: "utcodingbootcampproject1.appspot.com",
-    messagingSenderId: "316462943623"
-};
-firebase.initializeApp(config);
-
-
-
 // news apikey and query url set up
 var queryURL = 'https://newsapi.org/v2/top-headlines?' +
           'sources=fox-news&' +
